requirements.txt
requests
numpy
pandas
matplotlib
python-dotenv
pydanticstreamlit